database and UI support for selecting "executors"
An executor is something that executes the user code and tests.
Currently implemented (or half-implemented, really) are JUnit4 and
libcppunit; the UI for everything is there, but the libcppunit backend
isn't yet.

Unfortunately, this causes a database schema change; it doesn't work
in-place, and I don't have the energy to work out how to write upgrade
scripts.  It may be that this means that to deploy I need to give this a
new plugin name, or maybe there's a simple ALTER TABLE postgres
statement that we can use eventually?