implement JUnit4 XML parsing
This involves slightly more moving parts than were really desireable.
We need to tweak the security policy, to allow the Java process to write
the output file; we also need to mess around with the feedback
generation (and I've deliberately kept all the old code in for
backwards-compatibility).

And, to no-one's surprise, there's something in the system that's
removing newlines from the feedback, and something else that is
introducing <br /> tags, which means that stack traces in the feedback
look hilariously ugly.